requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
boto3==1.40.71
botocore==1.40.71
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
fastapi==0.119.1
fastapi-pagination==0.15.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio-ffmpeg==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
motor==3.7.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
slowapi==0.1.4
apscheduler==3.10.4